requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
aiohttp==3.9.1
pandas==2.1.4
openpyxl==3.1.2
pyyaml==6.0.1
fake-useragent==1.4.0
tenacity==8.2.3